Change behavior of the failing zephyr commits

+++ scripts/generate_workflow.py
@@ -26,13 +26,17 @@ def generate():
       GHA_SA: "gh-sa-gcp-distributed-job-buck"
       DEBIAN_FRONTEND: noninteractive
       TZ: Europe/Warsaw
+    outputs:
+      COMMIT_ALREADY_BUILT: ${{{{ steps.download-zephyr.outputs.COMMIT_ALREADY_BUILT }}}}
     steps:
     - uses: actions/checkout@v2
     - name: Prepare environment
       run: ./scripts/environment_prepare.sh
     - name: Download Zephyr
+      id: download-zephyr
       run: ./scripts/download_zephyr.sh
     - name: Pass Zephyr as artifact
+      if: steps.download-zephyr.outputs.COMMIT_ALREADY_BUILT == false
       run: |
         mkdir -p job-artifacts/prepare-zephyr-{zephyr_commit}
         mv zephyr.tar.gz job-artifacts/prepare-zephyr-{zephyr_commit}/
@@ -43,6 +47,7 @@ def generate():
     container: ubuntu:{UBUNTU_VERSION}
     runs-on: [self-hosted, Linux, X64]
     needs: [prepare-zephyr-{zephyr_commit}]
+    if: needs.prepare-zephyr-{zephyr_commit}.outputs.COMMIT_ALREADY_BUILT == false
     env:
       SAMPLE_NAME: {sample}
       MICROPYTHON_VERSION: 97a7cc243b
